requirements.txt
gitpython>=3.1.0
streamlit


